[feat] GroupCard.tsx 컴포넌트 분리
